pyproject.toml
[tool.poetry]
name = "jilm"
version = "0.1.0"
description = "Enhance your AI capabilities with jilm"
authors = ["Jean-Baptiste Poullet"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.154"
pygptj = "^1.0.10"
python-dotenv = "^1.0.0"
transformers = "^4.28.1"
sketch = "^0.3.5"
matplotlib = "^3.7.1"



[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
